requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.6
Django==3.2.8
django-address==0.2.5
django-crispy-forms==1.13.0
django-widget-tweaks==1.4.8
Pillow==8.3.2
pytz==2021.3
sqlparse==0.4.2
mysqlclient